docs/requirements.txt
GitPython
ipykernel
nbsphinx==0.8.8
pandoc
sphinx
sphinx_autodoc_typehints
sphinx_rtd_theme
